requirements.txt
ibm_db_sa
ibm_db
mysql-connector-python
python-dotenv